% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_.R
\name{read_cte501}
\alias{read_cte501}
\alias{read_mrfss}
\title{Read catch estimate and total mortality reports, i.e., 'CTE501'}
\usage{
read_cte501(file)

read_mrfss(file)
}
\arguments{
\item{file}{A file path to the file you want to work with.
The path can either be full or relative to your current working directory.}
}
\value{
A data frame.
}
\description{
Read in a MRFSS file containing catch estimates and total mortality reports.
Read in RecFIN data
}
\details{
Read in an excel file that contains information from RecFIN.
}
\section{cte501}{

Estimates of recent catches.
}

\section{mrfss}{

Estimates of catches from thes
Marine Recreational Fisheries Statistics Survey (MRFSS).
Catches are for the years between 1980 and approximately 2003
for all three states.
}

\author{
Kelli Faye Johnson
}
