% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recfin_coldefs}
\alias{recfin_coldefs}
\title{A data frame of column descriptions}
\format{
A data frame with four columns
\itemize{
\item Source: What data file does this pertain to, such as
MRFSS or cte50
\item Type: What data type within the source does this data align with,
for example in the MRFSS file there are multiple tabs of column
definitions and Type aligns with the tab name
\item Column: The column name in the data
\item Description: A description of what the column name means and
information about the data contained within the column
}
}
\usage{
recfin_coldefs
}
\description{
A data frame of column descriptions for standard files
provided by RecFIN.
}
\keyword{datasets}
