% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recfin_sexdefs}
\alias{recfin_sexdefs}
\title{A data frame of sex codes}
\format{
A data frame with seven columns
\itemize{
\item AGENCY_CODE: An integer specifying which agency the row pertains to
\item STSATE_NAME: The full character version of the state name in
title case
\item AGENCY_SEX_CODE: Integer or single character codes used by the agency
\item SEX_CODE: Upper-case character codes used by RecFIN
\item SEX_NAME: A description of what the SEX_CODE means
\item state: The two-letter abbreviation for the state name
\item code: An upper-case letter for the sex code provided by @kellijohnson-NOAA
with the following options: F, H, M, U, T
}
}
\usage{
recfin_sexdefs
}
\description{
A data frame of sex codes by state
provided by RecFIN.
}
\keyword{datasets}
